+++ src/Service/Lambda/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: Add Python3.15 (python3.15) and NodeJs 26 (nodejs26.x) support to AWS Lambda
+
 ## 2.17.0
 
 ### Added